Update Vermont data (#438)
## Description

I spent a while getting the Vermont sheet ready for the sheet-to-JSON
process. (As one of our earliest sheets, it's never been through this
process before.) Mainly:

- Adding necessary columns

- Fixing program names and making program URLs match the names

- Cutting down long descriptions

- Making per-filing-status and per-income-bracket copies of VT-27

- Excluding some incentives because they're expired, too vague, or not
  something we currently represent (the latter being the "replace my
  ride" ICE-car-scrapping incentives)

- Tweaking authority types (we're saying EVT is a state-level
  authority for these purposes) and adding the geo-eligibility column

- Adding a bunch of prereq and exclusion relationships, to express
  what's going on between utility-specific and EVT incentives.

I used `scripts/authority-and-program-updater.ts` and
`incentive-spreadsheet-to-json.ts`, and added the relationships
manually.

I'll follow up with a PR to add a bunch of EV incentives from utilities.

## Test Plan

`yarn test`. The differences in the fixture are commented; they are
significant.