Add Vermont utility EV incentives
## Description

BED, GMP, WEC, and VEC all have EV incentives. There are at least
three rows for each, as there are different amounts for new/used and
at least one income-qualified bonus for each.

One interesting note is that the VEC income-qualified bonus (VT-96) is
eligible if you qualify for the State of Vermont's higher
income-qualified bonus, so it has a prereq relationship on any of
those instead of a `low_income` field on itself. (That prereq is in
addition to requiring the base VEC incentive.)

As far as I can tell, you can claim all of these in addition to the
State's EV incentives, so there are no exclusion relationships needed.

## Test Plan

`yarn test`. `yarn check-spreadsheet-health` passes for VT.