Colorado updates from EnergySmart (#444)
## Description

This is all from updating the spreadsheet. The City of Boulder
incentives are now just flat amounts, not percents. A Boulder County
solar incentive and state-level tax credit for batteries are removed.

I updated the Spanish descriptions myself; the updates are
straightforward
and there are plenty of other examples to model them on.

https://app.asana.com/0/1206661332626418/1207127626720501/f

## Test Plan

`yarn test`