Add more items to the schema (#456)
## Description

This does not actually change any of the incentive data; just adds new
granular items to the schema.

I think this will be the best place to discuss the editorial matter of
exactly which items to add. Some of the decisions I made:

- I added `other_heat_pump` as a catchall for heat pump-like devices
  that don't fall under `ducted_heat_pump`, `ductless_heat_pump`,
  `central_air_conditioner`, `air_to_water_heat_pump`, or
  `geothermal_heating_installation`. I don't know if such a thing
  really exists, or if it does whether we care to include it in the API?

  - Note that my plan is to deprecate
`heat_pump_air_conditioner_heater`, rather than use it as the catchall.

- I also added (a) `other_insulation` and (b) `other_weatherization`.
  Respectively, these are for (a) insulation that isn't covered by one
  of the many granular categories I added; and (b) any weatherization
  measure that is not insulation and is not covered by one of the
  granular non-insulation measures. Similar to above, my plan is to
  deprecate `weatherization` rather than use it as the catchall.

- I added `{new,used}_plugin_hybrid_vehicle`. We've seen several
  incentives for thesespecifically, and I think it's worth splitting
  them out from `{new,used}_electric_vehicle`. The latter will become
  pure-EV only.

On the side, I made it so that human-friendly names and URLs are only
required for the items that appear in the IRA incentives (since
they're only needed in API v0).