Unlock calculation for territories
## Description

We can do this now that we have AMI data and 30C eligibility data for
the territories. This further required:

- Making the solar system cost estimation optional

- Making the estimated savings lookup fall back to the `US` value if
  it's not known for the user's state.

## Test Plan

Tested old and new embed frontends against local API server, with zip
00930 (PR).  The old one correctly shows total tax savings of $0 for
low and high incomes. The rooftop solar credit is shown as "30%"
instead of an estimated amount.

The tax credits still show up as eligible, which I think is not ideal,
but is OK -- territory residents are technically eligible for the
credits, and may benefit from them if they have non-territory-sourced
income or some other circumstance that means they do pay nonzero
federal income tax. (And this is the same thing you see for a low
income in the states.)