Update Spanish URLs for IRA incentives (#457)
## Description

We didn't update these upon launching the new consumer site content
because it wasn't in Spanish yet. Now it is.

Note that this affects v1 / the new calculator only. v0 and the old
embed still have to link to the `www` equipment pages (URLs in
`locales/{en,es}.json`) because those URLs are relative.

## Test Plan

`yarn test`. Visit each URL to make sure it works.