Store multiple items per incentive in JSON (#459)
## Description

The API is already serving an `items` array; now make the JSON data
store an `items` array and parse multiple values from the
corresponding spreadsheet column.

None of the spreadsheets have multiple values in that column yet, so
the data changes are purely mechanical at this point, not substantive.

I did the data changes by running the script in most cases (any state
for which the spreadsheet health check passes) and doing a find/replace
in others (CT, RI, NY, VA, IRA).

One thing to note is that I had to update any spreadsheet with electric
lawn equipment incentives, because the spreadsheet value for that
category contains commas, which breaks the comma-separation of multiple
values that this field is now doing. I changed the canonical value to
`Electric outdoor equipment`, and updated the incentives and enum value
sheets.

## Test Plan

`yarn test`.