Script and data updates for split items (#468)
## Description

I went through all the spreadsheets that currently pass the health
test, updated the item enums (on the enum values sheet) to include all
the new items, then updated all the rows that were about HVAC and
weatherization to have the more specific values.

There are plenty of cases where multiple items are needed in one
row. For these, I:

1. turned off the data validation in the spreadsheet,

2. manually set their "technology" column to comma-separated item
   values, such as `Door Replacement,Window Replacement`

3. turned data validation back on.

This is cheating to get around the validation, and it'll be annoying
to anyone who wants to update those cells, but I'm hoping the
spreadsheets won't last much longer.

In addition, I got the RI sheet passing the health test. So **RI is
the only data file that really needs an in-depth review**. I had to
add more program entries, and I made some substantive updates to the
incentives (other than just the items).

For other states where the health check is not passing, I just updated
the JSON.

### Follow-up

I need to make sure the frontend can handle all the extent multi-item
incentives. In particular, in my PR that added the concept of item
groups to the frontend, I didn't handle incentives that combine audit
and weatherization services.

## Test Plan

`yarn test`, and `yarn check-spreadsheet-health`. (The latter is
currently
failing because a bunch of Spanish translations just got added, but
that's the only diff.)