Exclude IRA rebates from v1
## Description

The embed is already filtering these out and replacing them with its
own content. I'd say PEP shouldn't be showing them at all, since
they're not actionable.

We do still need to keep them in the v0 API, to avoid disrupting the
legacy embed and v0 API clients.

The boolean flag is inelegant, but it will be removed along with the
v0 API.

## Test Plan

`yarn test` with new/updated tests.