Split items in IRA incentives
## Description

This splits up the combined items in the IRA incentives, allowing me
to get rid of `heat_pump_air_conditioner_heater` and `weatherization`
from the main items enum.

To keep the v0 API unchanged, I've added a new `v0_item` field to the
IRA incentive data. The values of that enum are no longer a subset of
the main item enum, because the two named above are still in
there. The field is allowed to be null, to exclude the incentive from
v0 altogether; I've done this with the new incentive for the energy
audit portion of the 25C tax credit. (`energy_audit` doesn't exist in
v0, and it doesn't make sense to have this part of the credit as a
separate `weatherization` incentive.)

## Test Plan

`yarn test`. In particular make sure the only change in the v0 fixture
is the changed short description.