Support per-county low income thresholds
## Description

This has been hanging over our heads since Vermont, but now NY's HEAR
program has them, so it's time to actually do it.

I added a `type` field to each threshold definition, which can have
the values `hhsize` or `county-hhsize`, indicating the sequence of
keys to be used with the `thresholds` structure. Counties are
identified by FIPS code, same as in the authorities files.

To demonstrate this working, I updated the county-specific Vermont
incentives (it's a lot less data than the NY ones, which I'll do
separately).

I've also switched the low-income-thresholds JSON-Schema-to-TS logic
away from ajv and to `json-schema-to-ts`. The latter is what I prefer
to use for this because it's what Fastify uses to turn endpoint
schemas into types, and I think it's worthwhile to have everything use
the same solution for that (because every schema-to-TS thing has
subtle differences). I've been converting things gradually. Anyway,
this resulted most notably in having to remove the `required:
[1,2,...]` constraint on the household-size-thresholds struct, so I
added a unit test for that.

## Test Plan

New test for county-specific Vermont thresholds (same income is LI in
one county but not in another). Other tests pass.