Refreshing expired CT data (#477)
Updated the source spreadsheet as well and created a named revision for
it.

For CT I checked things manually since there's not a lot of data and
since I wanted to see what this process would be like. It wasn't bad but
seems like it would be pretty awful for e.g. CO so I need to look at the
automation work that was done for this and maybe pick that back up.