Add NY's EmPower+; remove other NY incentives
## Description

We need to add this program to the calculator very soon, but the other
NY incentives we've gathered haven't been looked at in several months,
and they haven't been through the most up-to-date QA process at all.
Rather than rushing to QA them in time for EmPower's launch, I'm just
taking them out entirely.

They're coded as `efficiency_rebates` for expediency. This `item`
value is displayed by the frontend as "an energy efficiency retrofit",
which seems to be how EmPower+ is presented: it's a whole package deal.

The income thresholds came in the form of Javascript on the NYSERDA
eligibility page, so I wrote a quick script to populate them in the
low-income-thresholds file; that's included here. (There's one set of
thresholds per county in NY, so I didn't want to do it all manually.)

## Test Plan

Run the frontend against local API server; make sure it shows up
reasonably. (I posted an example here:
https://rewiringameri-g3x1100.slack.com/archives/C0711MM7K8R/p1718665502271849 )

For the script: run it and look at the resulting thresholds changes.
Spot-check a few of the FIPS codes to make sure it matches what the
NYSERDA webpage says if you use their calculator tool.