Replace update-fixtures.sh with env var in test itself (#490)
## Description

It's kinda silly to have to define every test case in two places like
this. This is a strategy I've used before: have an env var that
controls whether the test does the "snapshot == live" assertion, or
just overwrites the snapshot with the live value.

The workflow for adding a new test case is to write the
`validateResponse` call in v1.test.ts, including the query parameters
and the path for the new snapshot file, then run `yarn
update-snapshots` to generate the initial snapshot. Manually check to
make sure it's as expected.

For updating an existing test case, just change the query parameters
in the test case (if necessary) and run `yarn update-snapshots`. Check
to make sure the changes are as expected.

There's no way to selectively update individual snapshots; I can add
that if it proves necessary, but it might not.

While I'm at it, update the naming of these things to "snapshot"
rather than "fixture", as that's what it is.

## Test Plan

Make sure all the fixture/snapshot files are unchanged.

Go through the above workflows for adding a new test case and updating
an existing one. `yarn test` still passes.