Don't return ineligible incentives from v1
## Description

The fact that ineligible incentives were returned from `/calculator`
is a holdover from v0 that no longer makes sense. In v0, it was done
support showing ineligible IRA incentives grayed-out. But in the new
calculator, we're not doing that. Also, as a matter of API design,
it's conceptually unclean: it makes more sense if the endpoint just
returns a set of incentives that fulfill the criteria passed in.

Both embed and PEP clients are immediately filtering out ineligible
incentives anyway, so this doesn't affect them at all.

The next step will be to remove the `eligible` field from the API.

## Test Plan

`yarn test`. Look at the snapshot diffs and make sure all the removed
incentives have `eligible: false`. Grep the snapshots directory for
`"eligible": false` and make sure the only hit is in the v0 snapshot.