Refreshing data for VT (#493)
For VT, I updated the spreadsheet manually for anything that had an end
date in the past, and I also corrected a few inconsistencies between our
data and how the incentive programs were written. Then, I used the
spreadsheet to JSON script to update the JSON, which worked great and
definitely saved a bunch of time.

I'm starting to think about how to automate the data update part itself.
I haven't looked into whether the LLM data collection process is
deterministic when it comes to the ordering of the resulting incentives,
but I expect it's not. I think that makes it really hard to use it as-is
to refresh the data. Ideally we can have a mapping between old and new
incentives so we can only look at what changed instead of re-reviewing
all of the data after a refresh.