Remove Tri-State G&T incentives; fix some CO EV incentives
## Description

For background, Tri-State G&T is a cooperative made of a bunch of
rural/municipal electric utilities. They're an electricity wholesaler;
individual customers do not deal with them.

While collecting data for multiple Tri-State members, we noticed that
their rebate info was just the same Tri-State branded PDF. So we coded
those incentives as belonging to Tri-State, and created a geo group
for it.

However, the presentation of this was confusing. There's no Tri-State
hosted place for this program info (the PDF is hosted on each member's
site), so the URL had to go to United Power, which is a utility.

So instead, just remove the Tri-State versions of the incentives. Each
member utility that uses the generic Tri-State PDF already has a copy
of those incentives, so the Tri-State ones are redundant. It's also
not guaranteed that every Tri-State member offers those common
incentives, so the geo group representation is over-promising.

As part of this, doing some minor updates to La Plata Electric's EV
incentives, as pointed out to us by an employee. Adjusted the amount
on one. Removed one (which was basically a subset of another one).

## Test Plan

`yarn test`.