Launch NY (#518)
## Description

Only the EmPower+ program is in JSON now -- we've collected some other
incentives but that was a while ago and they haven't been QAed
recently. So to get EmPower+ out sooner, we're launching it by itself.

## Test Plan

`yarn test`. Hit local API with ZIP 12084 and low income and no
beta-states flag; make sure EmPower+ rebates are returned.