Add relationships to Clean Heat RI (#539)
## Description

This is a longstanding ticket that we somehow just never got around
to: having the income-qualified Clean Heat RI incentives exclude the
non-qualified ones.

34 (income-qualified ASHP + electrical upgrade) excludes 30
(non-qualified ASHP) and 32 (non-qualified electrical upgrade).

35 (income-qualified HPWH) excludes 33 (non-qualified HPWH).

RI-31 is a GSHP incentive which has no income-qualified equivalent, so
it's not excluded here.

https://app.asana.com/0/1206661332626418/1206661619324817

## Test Plan

`yarn test`. Snapshot tests cover this case. Query the API with a RI
zip code and low income; make sure the non-income-qualified versions
aren't returned.