New low-income threshold type for AMI percentage; use in WI
## Description

Focus on Energy's income-qualified incentives are now based on AMI,
rather than per-state thresholds like last year. Fortunately, we
already have up-to-date AMI data at hand, so just plumb it in to the
low-income calculation.

I confirmed that the numbers on their site match the numbers in our
AMI database in a few randomly selected counties.

## Test Plan

`yarn test` with new snapshot test (which fails on the old income thresholds).