Add optional "description" fields to a few types (#568)
## Description

Since none of the existing prereq relationships uses recursive
anyOf/allOf relationships, and it's super hard to support in a
relational format, I removed support for that. Now prereq
relationships are just `allOf: [list of strings]` or `anyOf: [list of
strings]` or just a single string.

Adding the field to exclusion relationships meant I had to add an
option to expand the list of IDs to an object, with an `ids` key (the
list of IDs) and the description.

## Test Plan

`yarn test`

Import from my local incentive admin with changes applied to export
descriptions in these positions, and make sure tests still pass here.