[NOT FOR MERGE] Import extra columns from spreadsheet
This will be the final step in spreadsheet deprecation.

The changes here are to the JSON schema and spreadsheet import
scripts, to bring in all the extra columns that are currently in
spreadsheets but nowhere else. The WI data is added here, as a
demonstration.

I don't plan to merge this to main, because there's no reason for
these extra columns to be in this codebase long-term. This is just the
most convenient way to get those columns into HERO. (The HERO change
to add these columns to the DB is already approved:
rewiringamerica/incentive_admin#79).

So the plan is:

1. Tell everyone to stop editing spreadsheets (forever) and HERO
   (temporarily).

2. Land the HERO PR to add the DB columns. Deploy to prod.

3. Run an export from HERO to here. Manually verify that there are no
   changes that exist only in HERO, then reset back to main.

3. Using the code in this PR, locally run an import from each
   spreadsheet currently in the `registry` file. That will bring the
   latest spreadsheet data, including the extra columns, into local
   JSON.

4. Get these changes reviewed, but don't merge them to main.

4. Import to prod HERO from local JSON. This is the last ever import.

5. Now people can edit HERO again.