update Efficiency Maine mobile home heat pump incentive (#585)
From EM: "One other inconsistency is on the heat pump offers. We do not
have a rebate for $2,000 for a ducted system. I think what is trying to
be expressed there is that we have a Manufactured (Mobile) Home
Initiative that pays the complete installation of a heat pump and
removal of the fossil system with a customer co-pay of $2,000, where
Efficiency Maine pays the remaining cost ($12,900)"