Foundation of gas utilities in API (#607)
## Description

- New authority type for gas utilities

- New key in the `/v1/utilities` response for gas utilities

- New query parameter in `/v1/calculator`

Currently, there are no gas utilities (I need to do some HERO-side
work for that) and the query param to `/calculator` doesn't do
anything. This is just laying the foundation so HERO can start
exporting gas utilities without causing schema tests to fail.

One thing I'm still undecided on is whether it's necessary to
distinguish in the API between "the choice of gas utility doesn't
matter in this case" and "there are no gas utilites serving this
location". I imagined indicating the former case by not having the
`gas_utilities` key in the `/v1/utilities` response, and the latter
case by having the key present and mapped to `{}`. But I'm not sure
it's necessary to make this distinction, or maybe if it should be made
another way (e.g. a boolean flag or something).

We still have time to decide that question, since this won't be
officially a part of the API for a bit.

## Test Plan

`yarn test`.

Manually add a `gas_utility` entry to one of the authorities.json
files; make sure `yarn test` still passes.