Fix how resolved location is returned in API (#616)
## Description

- The v1 APIs weren't returning the entire resolved location; the
  utilities endpoint was just returning state, and neither was
  returning county FIPS code.

- The API spec still included a `county` key, which we weren't
  returning under any circumstances. Change it to `countyFips`.

Note that mapping a zip code to a single city and a single county is
unsound -- ZCTAs cross city and county (and even state) boundaries all
the time. This PR isn't intended to rectify that, but just to make the
behavior match the currently-intended behavior.

## Test Plan

`yarn test` with updated tests.