Add MA gas utilities + some electric utility name changes (#619)
## Description

This is changes selected from a HERO import, not manual entry. I
removed a bunch of newly-added utilities because they don't have zip
codes defined, which causes a test to fail. (This test exists because
a utility that doesn't have zip codes defined is never going to show
up in `/utilities`, so it's useless. HERO doesn't currently have a way
to add zip codes to a utility, so this is a larger can of worms than I
want to deal with here.)

## Test Plan

`yarn test`