Add gas_utilities to geo groups; change multi-type matching logic
## Description

Mass Save is the first example of this type, containing both utilities
and gas_utilities.

The code said that the request has to match on _all_ of the group's
defined categories, but I don't remember why I wrote it that way, and
it doesn't really make sense to me now.

In any case, this isn't an observable behavior change because there
were no multi-type groups before.

## Test Plan

`yarn test`