Launch CT and OR; put IN and NM in beta
## Description

Had to update some tests that assumed CT was in beta; in doing so I
found a bug that had been around for a while. We would have been
underestimating eligibility for clients that passed
`authority_types=other`; offhand I would guess that this is zero
clients. (It was not affecting the calculator, because the calculator
doesn't pass the `authority_types` param.)

## Test Plan

`yarn test`