Slightly rework gas utilities API (#625)
## Description

This will make the API more suitable for general use, while still
supporting the calculator's intended behavior of only showing the gas
utility selector when necessary.

In addition to the recently added `gas_utilities` key in the
`/v1/utilities` response, I added `gas_utility_affects_incentives`.
If we don't have gas utility data for a state at all, both keys are
absent. If we have gas utility data, `gas_utilities` is always
populated and `gas_utility_affects_incentives` is present.

This way, clients who are just interested in which utilities serve the
area get whatever data we have. But the calculator can look at the
boolean flag to see whether it should show the selector.

## Test Plan

Testing is incomplete because we only have gas utility incentives in
MA, which is special-cased in the code anyway. But the new and
existing `/utilities` tests are decent coverage.