Filter out incentives with end dates in the past
## Description

Because of the more-than-just-a-day date formats we support, this is a
little bit subtle.

Note that this does not filter on incentive status, which we are not
exporting from HERO yet.

Note also that this uses @js-joda, a competent date/time library;
Javascript's native Date is a nightmare.

## Test Plan

Unit test of the new date logic. Snapshot tests reveal a bunch of
expired incentives that don't get returned anymore!