Filter out incentives with end dates in the past (#629)
## Description

Because of the more-than-just-a-day date formats we support, this is a
little bit subtle.

Note that this does not filter on incentive status, which we are not
exporting from HERO yet.

Note also that this uses js-joda, a competent date/time library;
Javascript's native Date is a nightmare.

## Test Plan

Unit test of the new date logic. Snapshot tests reveal a bunch of
expired incentives that don't get returned anymore!

Related, though: this makes the results of snapshot test runs
**time-dependent**, which is really not ideal. This will cause
test failures on completely unrelated PRs. The proper solution is to
inject the "current time" as a parameter to the filtering logic, instead
of calling `LocalDate.now()`, but that may have to wait; we want to
get this change out before the end of year break.