Check child zips for utility mappings as well (#658)
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209236041655647)

## Description

The ENERGY STAR electric utility dataset has a bunch of mappings to
zips that aren't ZCTAs. They aren't always consistent with the "parent
ZCTA" in our dataset. For example, in the case I added as a unit test:

- 01472 is not a ZCTA. Our dataset has 01450 as its parent ZCTA.
- `ma-groton-electric-light-dept` has a mapping to 01472, but _not_ to
01450.

It is clearly the case that Groton Electric Light covers the 01450
ZCTA (it encompasses the entire town of Groton), so arguably this is a
data error. Rather than go into a rabbit hole of trying to work out
the scope of the problem and update the zip/utility mappings to fix
it, I did this quick workaround to improve our results in these cases.

This change cannot decrease the set of results `/v1/utilities` will
return for any location, which is important.

## Test Plan

New test fails without the change.